+++ src/Actions/index.types.ts
@@ -5,15 +5,15 @@ export type SpinoramaActionsProps = BoxProps & {
 	/**
 	 * A reference to the next button element
 	 */
-	nextref?: React.Ref<HTMLButtonElement>
+	nextRef?: React.Ref<HTMLButtonElement>
 
 	/**
 	 * A reference to the prev button element
 	 */
-	prevref?: React.Ref<HTMLButtonElement>
+	prevRef?: React.Ref<HTMLButtonElement>
 
 	/**
 	 * A reference to the array of thumbnail button elements
 	 */
-	thumbnailsref?: React.MutableRefObject<(HTMLButtonElement | null)[]>
+	thumbnailsRef?: React.MutableRefObject<(HTMLButtonElement | null)[]>
 }