+++ src/Wrapper/index.types.ts
@@ -4,7 +4,7 @@ import type { ResponsiveStyleValue, GridSpacing } from '@mui/system'
 
 export type SpinoramaWrapperProps = BoxProps & {
 	/**
-	 * Determines how the slides are aligned.
+	 * Determines how the slides are aligned
 	 * Available options are:
 	 * - 'start': aligns the slides to the left (or top if `direction` is 'column');
 	 * - 'end': aligns the slides to the right (or bottom if `direction` is 'column');
@@ -14,16 +14,17 @@ export type SpinoramaWrapperProps = BoxProps & {
 	justify?: 'start' | 'end'
 
 	/**
-	 * Determines the direction of the slideshow.
-	 * When `true`, the slides will be displayed in reverse order.
+	 * Determines the direction of the slideshow
+	 * When `true`, the slides will be displayed in reverse order
 	 *
 	 * @default false
 	 */
 	reverse?: boolean
 
 	/**
-	 * Defines the space between the type `item` components.
-	 * It can only be used on a type `container` component.
+	 * Defines the space between the type `item` components
+	 * It can only be used on a type `container` component
+	 *
 	 * @default 0
 	 */
 	spacing?: ResponsiveStyleValue<GridSpacing>